18. emit
