+++ Juyoung/vue-project/src/App.vue
@@ -1,16 +1,18 @@
 <template>
   <div class="container">
     <h2>To-Do List</h2>
-    <TodoSimpleForm />
+    <TodoSimpleForm 
+      @add-todo="addTodo"
+    />
     <div v-if="!todos.length">
       추가된 Todo가 없습니다.
     </div>
     <div 
-      v-for="todo in todos" :key="todo.id"
+      v-for="(todo, index) in todos" :key="todo.id"
       class="card mt-2"
     >
-      <div class="card-body p-2">
-        <div class="form-check">
+      <div class="card-body p-2 d-flex align-items-center">
+        <div class="form-check flex-grow-1">
           <input 
             class="form-check-input" 
             type="checkbox"
@@ -22,6 +24,12 @@
           >
             {{ todo.subject }}
           </label>
+        </div>
+        <div>
+          <button 
+            class="btn btn-danger btn-sm"
+            @click="deleteTodo(index)"
+          >Delete</button>
         </div>
       </div>
     </div>
@@ -37,34 +45,24 @@ export default {
     TodoSimpleForm
   },
   setup() {
-    const todo = ref('');
     const todos = ref([]);
-    const hasError = ref(false);
     const todoStyle = {
       textDecoration: 'line-through',
       color: 'gray'
     }
+    const addTodo = (todo) => {
+      todos.value.push(todo);
+    }
 
-    const onSubmit = () => {
-      if(todo.value === '') {
-        hasError.value = true;
-      } else {
-        todos.value.push({
-          id: Date.now(),
-          subject: todo.value,
-          completed: false,
-        });
-        hasError.value = false;
-        todo.value='';
-      }
+    const deleteTodo = (index) => {
+      todos.value.splice(index, 1);
     }
 
     return {
-      todo,
-      onSubmit,
+      addTodo,
       todos,
-      hasError,
-      todoStyle
+      todoStyle,
+      deleteTodo
     };
   }
 }
