+++ Juyoung/vue-project/src/components/TodoSimpleForm.vue
@@ -18,14 +18,38 @@
           </button>
         </div>
       </div>
-      <div v-show="hasError" style="color: red">This field cannot be empty</div>
+      <div v-show="hasError" style="color:red">This field cannot be empty</div>
     </form>
 </template>
 
 <script>
-export default {
+  import { ref } from 'vue';
+  export default {
+    setup(props, context) {
+      const todo = ref('');
+      const hasError = ref(false);
+      const onSubmit = () => {
+        if(todo.value === '') {
+          hasError.value = true;
+        } else {
+          context.emit('add-todo', {
+            id: Date.now(),
+            subject: todo.value,
+            completed: false,
+          });
+          hasError.value = false;
+          todo.value='';
+        }
+      }
 
-}
+      return {
+        todo,
+        hasError,
+        onSubmit
+      }
+    }
+
+  }
 </script>
 
 <style>
